added action_bundle_test.go
