+++ action.go
@@ -5,7 +5,6 @@
 package holochain
 
 import (
-	"bytes"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -13,8 +12,6 @@ import (
 	b58 "github.com/jbenet/go-base58"
 	ic "github.com/libp2p/go-libp2p-crypto"
 	peer "github.com/libp2p/go-libp2p-peer"
-	"io/ioutil"
-	"net/http"
 	"reflect"
 	"time"
 )
